#6: Rename render test util file
