+++ packages/webapp/src/app/__tests__/page.test.tsx
@@ -1,5 +1,5 @@
 import Home from '../page';
-import renderTestWithProviders from '@/utils/test/renderWithProviders';
+import renderTestWithProviders from '@/utils/test/renderTestWithProviders';
 
 test('test', () => {
   const { container } = renderTestWithProviders(<Home />);
